refactor(beautiful-mermaid): extract runtime
