+++ packages/extensions/beautiful-mermaid/src/index.ts
@@ -1,4 +1,5 @@
 export * from './cdn'
 export * from './constants'
 export * from './extension'
+export * from './runtime'
 export * from './types'
